# frontend/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Route 53 Clone API",
    "description": "\nBackend of the Route 53 console clone. Hosted zones and record sets follow Route 53's\nrules: every zone owns an apex NS and SOA record, CNAMEs cannot share a name with other\nrecords, and change batches are atomic.\n\nSign in with `POST /api/v1/auth/login`; the session travels in an httpOnly cookie.\nErrors always have the shape `{code, message, details}`.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/hostedzones/{zone_id}": {
      "get": {
        "tags": [
          "Hosted zones"
        ],
        "summary": "Get a hosted zone",
        "operationId": "get_hosted_zone",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostedZoneDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Hosted zones"
        ],
        "summary": "Edit a hosted zone",
        "description": "The description, the tags and, for a private zone, the VPC associations can be changed after a zone is created. Omitted fields keep their value; the edit is applied as a whole or not at all.",
        "operationId": "update_hosted_zone",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostedZoneUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostedZoneDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request violates a Route 53 rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request body or parameters are malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Hosted zones"
        ],
        "summary": "Delete a hosted zone",
        "description": "Fails with `HostedZoneNotEmpty` while the zone has records besides the apex NS and SOA.",
        "operationId": "delete_hosted_zone",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The change conflicts with existing data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/hostedzones/{zone_id}/records": {
      "get": {
        "tags": [
          "Records"
        ],
        "summary": "List the records of a hosted zone",
        "operationId": "list_records",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 255
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text search across the main columns",
              "title": "Search"
            },
            "description": "Free-text search across the main columns"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/RecordType"
              },
              "description": "Only these record types; repeatable",
              "default": [],
              "title": "Type"
            },
            "description": "Only these record types; repeatable"
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Property filter as `field:operator:value`, repeatable. Operators: eq, ne, contains, not_contains, starts_with, not_starts_with, gt, gte, lt, lte.",
              "examples": [
                "name:contains:shop"
              ],
              "default": [],
              "title": "Filter"
            },
            "description": "Property filter as `field:operator:value`, repeatable. Operators: eq, ne, contains, not_contains, starts_with, not_starts_with, gt, gte, lt, lte."
          },
          {
            "name": "filter_mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "and",
                "or"
              ],
              "type": "string",
              "description": "Whether all filters must match or any of them",
              "default": "and",
              "title": "Filter Mode"
            },
            "description": "Whether all filters must match or any of them"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "default (Route 53 order), name, type, ttl, routing_policy, differentiator, set_identifier, alias, value, health_check_id, evaluate_target_health or id",
              "default": "default",
              "title": "Sort"
            },
            "description": "default (Route 53 order), name, type, ttl, routing_policy, differentiator, set_identifier, alias, value, health_check_id, evaluate_target_health or id"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "description": "Sort direction",
              "default": "asc",
              "title": "Order"
            },
            "description": "Sort direction"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Items per page",
              "default": 50,
              "title": "Page Size"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RecordSetOut_"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone or record does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request violates a Route 53 rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request body or parameters are malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Records"
        ],
        "summary": "Create a record",
        "operationId": "create_record",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordSetInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordSetOut"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone or record does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request violates a Route 53 rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request body or parameters are malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The change conflicts with existing data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hostedzones/{zone_id}/records:batch": {
      "post": {
        "tags": [
          "Records"
        ],
        "summary": "Apply a change batch atomically",
        "description": "CREATE, UPSERT and DELETE changes run in order in one transaction, like `ChangeResourceRecordSets`. If any change is invalid, none is applied and every failure is listed in `details` with the index of its change.",
        "operationId": "change_records",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeBatchResult"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone or record does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request violates a Route 53 rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request body or parameters are malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hostedzones/{zone_id}/records/{record_id}": {
      "get": {
        "tags": [
          "Records"
        ],
        "summary": "Get a record",
        "operationId": "get_record",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Record ID",
              "title": "Record Id"
            },
            "description": "Record ID"
          },
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordSetOut"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone or record does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Records"
        ],
        "summary": "Edit a record",
        "description": "Omitted fields keep their value. The apex NS and SOA records can be edited but not renamed or retyped.",
        "operationId": "update_record",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Record ID",
              "title": "Record Id"
            },
            "description": "Record ID"
          },
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordSetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordSetOut"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone or record does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request violates a Route 53 rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request body or parameters are malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The change conflicts with existing data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Records"
        ],
        "summary": "Delete a record",
        "description": "The apex NS and SOA records cannot be deleted.",
        "operationId": "delete_record",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Record ID",
              "title": "Record Id"
            },
            "description": "Record ID"
          },
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone or record does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request violates a Route 53 rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request body or parameters are malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeBatchRequest": {
        "properties": {
          "comment": {
            "type": "string",
            "maxLength": 256,
            "title": "Comment",
            "default": ""
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/Change"
            },
            "type": "array",
            "maxItems": 1000,
            "minItems": 1,
            "title": "Changes"
          }
        },
        "type": "object",
        "required": [
          "changes"
        ],
        "title": "ChangeBatchRequest"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "NoSuchHostedZone"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HostedZoneDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "Z0123456789ABCDEFGHIJ"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Canonical zone name with trailing dot",
            "examples": [
              "example.com."
            ]
          },
          "type": {
            "$ref": "#/components/schemas/ZoneType"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "record_count": {
            "type": "integer",
            "title": "Record Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "caller_reference": {
            "type": "string",
            "title": "Caller Reference"
          },
          "name_servers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Name Servers"
          },
          "vpcs": {
            "items": {
              "$ref": "#/components/schemas/VpcAssociation"
            },
            "type": "array",
            "title": "Vpcs"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "description",
          "created_by",
          "record_count",
          "created_at",
          "updated_at",
          "caller_reference",
          "name_servers",
          "vpcs",
          "tags"
        ],
        "title": "HostedZoneDetail"
      },
...
      "HostedZoneUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Tag"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "Replaces every tag of the zone"
          },
          "vpcs": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/VpcAssociation"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vpcs",
            "description": "Replaces the VPC associations of a private zone"
          }
        },
        "type": "object",
        "title": "HostedZoneUpdate",
        "description": "Partial update; omitted fields keep their current value."
      },
...
      "RecordType": {
        "type": "string",
        "enum": [
          "A",
          "AAAA",
          "CAA",
          "CNAME",
          "MX",
          "NS",
          "PTR",
          "SOA",
          "SRV",
          "TXT"
        ],
        "title": "RecordType"
      },
...
      "ZoneType": {
        "type": "string",
        "enum": [
          "public",
          "private"
        ],
        "title": "ZoneType"
      }
    }
  }
}